Remove LocalStorage fallback as this is senseless
We need more than ~2-5MB storage size, so a fallback to localstorage may cause
smt like this in your error logs:

NS_ERROR_DOM_QUOTA_REACHED: Persistent storage maximum size reached

This is because of the fallback to localstorage and the size is bigger than the
browsers available capacity, so an exception will be thrown.

So do we need to support a fallback to localstorage at all? answer: no.

Some notes on this
==================
* in cordova based environments we will have websql support for sure, nice.
* if a browser is in private mode there can be no offline support, fact.
* fallback to network if both storage modules fail is ok